refactor: put types in types.py for more clarity
